biomaRt-deprecated {biomaRt}	R Documentation

Deprecated functions in package ‘biomaRt’

Description

These functions are provided for compatibility with older versions of ‘biomaRt’ only, and will be defunct at the next release.

Details

The following functions are deprecated and will be made defunct; use the replacement indicated below:

filterOptions: listFilterOptions

listFilterValues: listFilterOptions

searchFilterValues: searchFilterOptions

The following functions are defunct and no longer work; use the replacement indicated below:

getBMlist: getBM

[Package biomaRt version 2.50.3 Index]