useDataset {biomaRt}	R Documentation

Select a dataset to use and updates Mart object

Description

This function selects a dataset and updates the Mart object

Usage

useDataset(dataset,mart, verbose = FALSE)

Arguments

dataset
Dataset you want to use. List of possible datasets can be retrieved using the function listDatasets

mart
Mart object created with the useMart function

verbose
Give detailed output of what the method is doing, for debugging

Author(s)

Steffen Durinck

Examples


if(interactive()){
mart=useMart("ensembl")
mart=useDataset("hsapiens_gene_ensembl", mart = mart)
}

[Package biomaRt version 2.38.0 Index]