listEnsembl {biomaRt}	R Documentation

lists the available BioMart databases hosted by Ensembl

Description

This function returns a list of BioMart databases hosted by Ensembl. To establish a connection use the useMart function.

Usage

listEnsembl(mart = NULL,host = "www.ensembl.org", version = NULL, 
                   GRCh = NULL, mirror = NULL, verbose = FALSE)

Arguments

mart
mart object created with the useEnsembl function. This is optional, as you usually use listMarts to see which marts there are to connect to.

host
Host to connect to if different then www.ensembl.org

version
Ensembl version to connect to when wanting to connect to an archived Ensembl version

GRCh
GRCh version to connect to if not the current GRCh38, currently this can only be 37

mirror
Specify an Ensembl mirror to connect to. The valid options here are 'www', 'uswest', 'useast', 'asia'. If no mirror is specified the primary site at www.ensembl.org will be used.

verbose
Give detailed output of what the method is doing, for debugging purposes

Author(s)

Steffen Durinck

Examples


if(interactive()){
listEnsembl()
}

[Package biomaRt version 2.38.0 Index]