listEnsemblArchives {biomaRt}	R Documentation

Lists the available archived versions of Ensembl

Description

Returns a table containing the available archived versions of Ensembl, along with the dates they were created and the URL used to access them.

Usage

listEnsemblArchives(https)

Arguments

https
Deprecated argument. Ensembl are enforcing https use from late 2021 and this argument will be removed at this time as it no longer serves a purpose. Originally - "Logical value of length 1. Determines whether https should be used to contact the Ensembl server."

Author(s)

Mike Smith

Examples

listEnsemblArchives()

[Package biomaRt version 2.50.3 Index]