listEnsemblArchives {biomaRt}	R Documentation

Lists the available archived versions of Ensembl

Description

Returns a table containing the available archived versions of Ensembl, along with the dates they were created and the URL used to access them.

Usage

listEnsemblArchives(https = TRUE)

Arguments

https
Logical value of length 1. Determines whether https should be used to contact the Ensembl server. Defaults to TRUE.

Author(s)

Mike Smith

Examples

listEnsemblArchives()

[Package biomaRt version 2.46.3 Index]