listEnsemblArchives {biomaRt}	R Documentation

Lists the available archived versions of Ensembl

Description

Returns a table containing the available archived versions of Ensembl, along with the dates they were created and the URL used to access them.

Usage

listEnsemblArchives()

Author(s)

Mike Smith

Examples


listEnsemblArchives()

[Package biomaRt version 2.38.0 Index]