useMart {biomaRt}	R Documentation

Connects to the selected BioMart database and dataset

Description

A first step in using the biomaRt package is to select a BioMart database and dataset to use. The useMart function enables one to connect to a specified BioMart database and dataset within this database. To know which BioMart databases are available see the listMarts function. To know which datasets are available within a BioMart database, first select the BioMart database using useMart and then use the listDatasets function on the selected BioMart, see listDatasets function.

Usage

useMart(biomart, dataset, host="www.ensembl.org",
path="/biomart/martservice", port=80, archive=FALSE, ssl.verifypeer =
TRUE, ensemblRedirect = NULL, version, verbose = FALSE)

Arguments

biomart
BioMart database name you want to connect to. Possible database names can be retrieved with the functio listMarts

dataset
Dataset you want to use. To see the different datasets available within a biomaRt you can e.g. do: mart = useMart('ensembl'), followed by listDatasets(mart).

host
Host to connect to. Defaults to www.ensembl.org

path
Path that should be pasted after to host to get access to the web service URL

port
port to connect to, will be pasted between host and path

archive
Boolean to indicate if you want to access archived versions of BioMart databases. Note that this argument is now deprecated and will be removed in the future. A better alternative is to leave archive = FALSE and to specify the url of the archived BioMart you want to access. For Ensembl you can view the list of archives using listEnsemblArchives

ssl.verifypeer
Set SSL peer verification on or off. By default ssl.verifypeer is set to TRUE

ensemblRedirect
This argument has now been deprecated.

version
Use version name instead of biomart name to specify which BioMart you want to use

verbose
Give detailed output of what the method is doing while in use, for debugging

Author(s)

Steffen Durinck

Examples


if(interactive()){

mart = useMart("ensembl")
mart=useMart(biomart="ensembl", dataset="hsapiens_gene_ensembl")
}

[Package biomaRt version 2.38.0 Index]